feat: migrate electron main file to ESM and update package.json type
