+++ package.json
@@ -102,5 +102,6 @@
       "jsdom": "25.0.1"
     },
     "neverBuiltDependencies": []
-  }
+  },
+  "type": "module"
 }
